Create a tabular representation of man file information

Provides Rd index info with a few additional columns of information about each exported object. Returns one record per documented object, even if multiple objects alias to the same documentation file.

Rd_df(x)

Arguments

x
A package object to coerce

Value

A data.frame of documented object information with variables:

index

A numeric index of documentation files associated with documentation objects

file

A character filename of the Rd file in the "man" directory

filepath

A character file path of the Rd file in the "man" directory

alias

character object names which are aliases for the documentation in filepath

is_exported

A logical indicator of whether the aliased object is exported from the package namespace

doctype

A character representing the Rd docType field.

Examples

package_source_dir <- system.file("examplepkg", package = "covtracer")
Rd_df(package_source_dir)

#>                          file
#> 1              Accumulator.Rd
#> 2               hypotenuse.Rd
#> 3   names-S4Example-method.Rd
#> 4  names-S4Example2-method.Rd
#> 5                   Person.Rd
#> 6                    Rando.Rd
#> 7          rd_data_sampler.Rd
#> 8               rd_sampler.Rd
#> 9          s3_example_func.Rd
#> 10         s3_example_func.Rd
#> 11         s3_example_func.Rd
#> 12         S4Example-class.Rd
#> 13         S4Example-class.Rd
#> 14        S4Example2-class.Rd
#> 15        S4Example2-class.Rd
#>                                                                               filepath
#> 1              /home/runner/work/_temp/Library/covtracer/examplepkg/man/Accumulator.Rd
#> 2               /home/runner/work/_temp/Library/covtracer/examplepkg/man/hypotenuse.Rd
#> 3   /home/runner/work/_temp/Library/covtracer/examplepkg/man/names-S4Example-method.Rd
#> 4  /home/runner/work/_temp/Library/covtracer/examplepkg/man/names-S4Example2-method.Rd
#> 5                   /home/runner/work/_temp/Library/covtracer/examplepkg/man/Person.Rd
#> 6                    /home/runner/work/_temp/Library/covtracer/examplepkg/man/Rando.Rd
#> 7          /home/runner/work/_temp/Library/covtracer/examplepkg/man/rd_data_sampler.Rd
#> 8               /home/runner/work/_temp/Library/covtracer/examplepkg/man/rd_sampler.Rd
#> 9          /home/runner/work/_temp/Library/covtracer/examplepkg/man/s3_example_func.Rd
#> 10         /home/runner/work/_temp/Library/covtracer/examplepkg/man/s3_example_func.Rd
#> 11         /home/runner/work/_temp/Library/covtracer/examplepkg/man/s3_example_func.Rd
#> 12         /home/runner/work/_temp/Library/covtracer/examplepkg/man/S4Example-class.Rd
#> 13         /home/runner/work/_temp/Library/covtracer/examplepkg/man/S4Example-class.Rd
#> 14        /home/runner/work/_temp/Library/covtracer/examplepkg/man/S4Example2-class.Rd
#> 15        /home/runner/work/_temp/Library/covtracer/examplepkg/man/S4Example2-class.Rd
#>                      alias is_exported doctype
#> 1              Accumulator        TRUE    <NA>
#> 2               hypotenuse        TRUE    <NA>
#> 3   names,S4Example-method       FALSE    <NA>
#> 4  names,S4Example2-method       FALSE    <NA>
#> 5                   Person        TRUE    <NA>
#> 6                    Rando        TRUE    <NA>
#> 7          rd_data_sampler        TRUE    data
#> 8               rd_sampler        TRUE    <NA>
#> 9          s3_example_func       FALSE    <NA>
#> 10 s3_example_func.default       FALSE    <NA>
#> 11    s3_example_func.list       FALSE    <NA>
#> 12         S4Example-class       FALSE   class
#> 13               S4Example       FALSE   class
#> 14        S4Example2-class       FALSE   class
#> 15              S4Example2       FALSE   class