Extract all the srcref objects of objects within a package namespace

Extract all the srcref objects of objects within a package namespace

pkg_srcrefs(x)

# S3 method for environment
pkg_srcrefs(x)

# S3 method for character
pkg_srcrefs(x)

# S3 method for coverage
pkg_srcrefs(x)

Arguments

x
A package_coverage coverage object, from which the name of the package used is extracted.

See also

Examples

pkg <- system.file("examplepkg", package = "covtracer")
remotes::install_local(
  pkg, 
  force = TRUE, 
  quiet = TRUE, 
  INSTALL_opts = "--with-keep.source"
)

#> Running `R CMD build`...

pkg_srcrefs("examplepkg")

#> $s3_example_func.default
#>  14
#>  28
#>  16
#>   1
#>  28
#>   1
#> 201
#> 203
#> 
#> $Rando
#>  87
#>  12
#>  94
#>   3
#>  12
#>   3
#>  99
#> 106
#> 
#> $Person
#> 52
#> 18
#> 56
#>  5
#> 18
#>  5
#> 64
#> 68
#> 
#> $Person
#> 64
#> 13
#> 69
#>  5
#> 13
#>  5
#> 76
#> 81
#> 
#> $`names,S4Example-method`
#>  15
#>  44
#>  17
#>   1
#>  44
#>   1
#> 224
#> 226
#> 
#> $`names,S4Example2-method`
#>  32
#>  45
#>  34
#>   1
#>  45
#>   1
#> 241
#> 243
#> 
#> $hypotenuse
#>  7
#> 15
#>  9
#>  1
#> 15
#>  1
#>  9
#> 11
#> 
#> $Accumulator
#> 17
#> 16
#> 20
#>  3
#> 16
#>  3
#> 29
#> 32
#> 
#> $Accumulator
#> 28
#>  9
#> 32
#>  3
#>  9
#>  3
#> 40
#> 44
#> 
#> $s3_example_func
#>   9
#>  20
#>  11
#>   1
#>  20
#>   1
#> 196
#> 198
#> 
#> $s3_example_func.list
#>  19
#>  25
#>  21
#>   1
#>  25
#>   1
#> 206
#> 208
#> 
#> $rd_sampler
#>  56
#>  15
#>  58
#>   1
#>  15
#>   1
#> 184
#> 186
#>