Extract all the srcref objects of objects within a package namespace

Extract all the srcref objects of objects within a package namespace

Usage

pkg_srcrefs(x)

# S3 method for environment
pkg_srcrefs(x)

# S3 method for character
pkg_srcrefs(x)

# S3 method for coverage
pkg_srcrefs(x)

Arguments

x

A package_coverage coverage object, from which the name of the package used is extracted.

See also

Examples

pkg <- system.file("examplepkg", package = "covtracer")
remotes::install_local(
  pkg,
  force = TRUE,
  quiet = TRUE,
  INSTALL_opts = "--with-keep.source"
)
#> Running `R CMD build`...
pkg_srcrefs("examplepkg")
#> $s3_example_func.default
#>  15
#>  28
#>  17
#>   1
#>  28
#>   1
#> 259
#> 261
#> 
#> $Rando
#>  95
#>  12
#> 102
#>   3
#>  12
#>   3
#> 132
#> 139
#> 
#> $adder
#>  3
#> 10
#>  9
#>  1
#> 10
#>  1
#> 40
#> 46
#> 
#> $Person
#>  60
#>  18
#>  64
#>   5
#>  18
#>   5
#>  97
#> 101
#> 
#> $Person
#>  72
#>  13
#>  77
#>   5
#>  13
#>   5
#> 109
#> 114
#> 
#> $`names,S4Example-method`
#>  16
#>  44
#>  18
#>   1
#>  44
#>   1
#> 283
#> 285
#> 
#> $`names,S4Example2-method`
#>  34
#>  45
#>  36
#>   1
#>  45
#>   1
#> 301
#> 303
#> 
#> $hypotenuse
#>  7
#> 15
#>  9
#>  1
#> 15
#>  1
#> 34
#> 36
#> 
#> $Accumulator
#> 29
#> 16
#> 32
#>  3
#> 16
#>  3
#> 66
#> 69
#> 
#> $Accumulator
#>  4
#>  3
#>  8
#>  3
#>  3
#>  3
#> 41
#> 45
#> 
#> $s3_example_func
#>  10
#>  20
#>  12
#>   1
#>  20
#>   1
#> 254
#> 256
#> 
#> $complex_call_stack
#>  3
#> 23
#>  5
#>  1
#> 23
#>  1
#>  5
#>  7
#> 
#> $recursive_function
#> 21
#> 23
#> 24
#>  1
#> 23
#>  1
#> 23
#> 26
#> 
#> $nested_function
#>  9
#> 20
#> 11
#>  1
#> 20
#>  1
#> 11
#> 13
#> 
#> $deeper_nested_function
#> 15
#> 27
#> 17
#>  1
#> 27
#>  1
#> 17
#> 19
#> 
#> $s3_example_func.list
#>  20
#>  25
#>  22
#>   1
#>  25
#>   1
#> 264
#> 266
#> 
#> $rd_sampler
#>  56
#>  15
#>  58
#>   1
#>  15
#>   1
#> 224
#> 226
#> 
#> $help
#> NA
#> 
#> $rd_data_sampler
#> NA
#> 
#> $person
#> NA
#> 
#> $reexport_example
#> NA
#> 
#> $PersonPrime
#> NA
#>