Create a data.frame of package srcref objects

Create a data.frame of package srcref objects

Usage

pkg_srcrefs_df(x)

Arguments

x

A package_coverage coverage object, from which the name of the package used is extracted.

Value

A data.frame with a record for each source code block with variables:

name

A character Rd alias for the package object

srcref

The srcref of the associated package source code

See also

Examples

pkg <- system.file("examplepkg", package = "covtracer")
remotes::install_local(
  pkg,
  force = TRUE,
  quiet = TRUE,
  INSTALL_opts = "--with-keep.source"
)
#> Running `R CMD build`...
pkg_srcrefs_df("examplepkg")
#>                       name                          srcref  namespace
#> 1  s3_example_func.default         s3_example.R:15:28:17:1 examplepkg
#> 2                    Rando        r6_example.R:95:12:102:3 examplepkg
#> 3                    adder           r6_example.R:3:10:9:1 examplepkg
#> 4                   Person         r6_example.R:60:18:64:5 examplepkg
#> 5                   Person         r6_example.R:72:13:77:5 examplepkg
#> 6   names,S4Example-method         s4_example.R:16:44:18:1 examplepkg
#> 7  names,S4Example2-method         s4_example.R:34:45:36:1 examplepkg
#> 8               hypotenuse           hypotenuse.R:7:15:9:1 examplepkg
#> 9              Accumulator         r6_example.R:29:16:32:3 examplepkg
#> 11         s3_example_func         s3_example.R:10:20:12:1 examplepkg
#> 12      complex_call_stack   complex_call_stack.R:3:23:5:1 examplepkg
#> 13      recursive_function complex_call_stack.R:21:23:24:1 examplepkg
#> 14         nested_function  complex_call_stack.R:9:20:11:1 examplepkg
#> 15  deeper_nested_function complex_call_stack.R:15:27:17:1 examplepkg
#> 16    s3_example_func.list         s3_example.R:20:25:22:1 examplepkg
#> 17              rd_sampler         rd_sampler.R:56:15:58:1 examplepkg
#> 18                    help                            <NA>      utils
#> 19         rd_data_sampler                            <NA>       <NA>
#> 20                  person                            <NA>      utils
#> 21        reexport_example                            <NA>      utils
#> 22             PersonPrime                            <NA>       <NA>