Create a data.frame of coverage test srcref objects

Extract unit test srcrefs from a coverage object. A test name will be derived from the test source code, preferrably from a written annotation, but otherwise falling back to using a code snippet. srcrefs are unique for each expression executed within a testing suite.

test_srcrefs_df(x)

Arguments

x
A package_coverage coverage object, from which the name of the package used is extracted.

Value

A data.frame of test srcrefs extracted from a coverage object. Contains one record for each srcref with variables:

name

A character test description. For testthat tests, the desc parameter will be used, otherwise a snippet of code will be used for the test name

srcref

A srcref object describing the location of the test

See also

Examples

options(covr.record_tests = TRUE)
pkg_path <- system.file("examplepkg", package = "covtracer")
cov <- covr::package_coverage(pkg_path)
test_srcrefs_df(cov)

#>                                                         name
#> 1                         hypotenuse is calculated correctly
#> 2  hypotenuse is calculated correctly; with negative lengths
#> 3         Example R6 Accumulator class constructor is traced
#> 4            Example R6 Accumulator class methods are traced
#> 5            Example R6 Accumulator class methods are traced
#> 6          Example R6 Person class public methods are traced
#> 7          Example R6 Person class public methods are traced
#> 8   Example R6 Rando class active field functions are traced
#> 9               s3_example_func works using default dispatch
#> 10                 s3_example_func works using list dispatch
#> 11                              S4Example names method works
#>                          srcref
#> 1    test-hypotenuse.R:2:3:2:35
#> 2    test-hypotenuse.R:5:5:5:39
#> 3    test-r6-example.R:2:3:2:43
#> 4    test-r6-example.R:7:3:7:43
#> 5    test-r6-example.R:8:3:8:36
#> 6  test-r6-example.R:12:3:12:40
#> 7  test-r6-example.R:13:3:13:26
#> 8  test-r6-example.R:18:3:18:38
#> 9    test-s3-example.R:2:3:2:46
#> 10   test-s3-example.R:6:3:6:54
#> 11   test-s4-example.R:3:3:3:40