Create a data.frame mapping tests to coverage traces

Extract a matrix used to relate test code to the traces that each test evaluates.

Usage

test_trace_mapping(x)

Arguments

x

A coverage object produced with options(covr.record_tests = TRUE).

Value

A data.frame with one record for each line of code executed, with variables:

test

The index of the test that was executed, reflecting the order in which tests are executed

depth

The call stack depth when the coverage trace was evaluated

i

The index of the expression evaluated by each test. This can be used to recover an order of trace execution for a given test index

trace

The index of the coverage trace that was evaluated

See also

Examples

options(covr.record_tests = TRUE)
pkg_path <- system.file("examplepkg", package = "covtracer")
cov <- covr::package_coverage(pkg_path)
test_trace_mapping(cov)
#>        test call depth  i trace
#>   [1,]    1    1     1  1     9
#>   [2,]    2    1    37  1    18
#>   [3,]    2    1    38  2    14
#>   [4,]    2    1    39  3    24
#>   [5,]    2    1    40  4    21
#>   [6,]    2    1    40  5    27
#>   [7,]    2    1    41  6    21
#>   [8,]    2    1    41  7    27
#>   [9,]    2    1    42  8    21
#>  [10,]    2    1    42  9    27
#>  [11,]    2    1    43 10    21
#>  [12,]    2    1    43 11    25
#>  [13,]    3    1    37  1    24
#>  [14,]    3    1    38  2    21
#>  [15,]    3    1    38  3    27
#>  [16,]    3    1    39  4    21
#>  [17,]    3    1    39  5    27
#>  [18,]    3    1    40  6    21
#>  [19,]    3    1    40  7    27
#>  [20,]    3    1    41  8    21
#>  [21,]    3    1    41  9    25
#>  [22,]    4    1    42  1    18
#>  [23,]    4    2    42  1    18
#>  [24,]    4    3    42  1    18
#>  [25,]    4    4    42  1    18
#>  [26,]    4    5    42  1    18
#>  [27,]    4    6    42  1    18
#>  [28,]    4    7    42  1    18
#>  [29,]    4    8    42  1    18
#>  [30,]    4    9    42  1    18
#>  [31,]    4   10    42  1    18
#>  [32,]    4    1    43  2    14
#>  [33,]    4    2    43  2    14
#>  [34,]    4    3    43  2    14
#>  [35,]    4    4    43  2    14
#>  [36,]    4    5    43  2    14
#>  [37,]    4    6    43  2    14
#>  [38,]    4    7    43  2    14
#>  [39,]    4    8    43  2    14
#>  [40,]    4    9    43  2    14
#>  [41,]    4   10    43  2    14
#>  [42,]    4    1    44  3    24
#>  [43,]    4    2    44  3    24
#>  [44,]    4    3    44  3    24
#>  [45,]    4    4    44  3    24
#>  [46,]    4    5    44  3    24
#>  [47,]    4    6    44  3    24
#>  [48,]    4    7    44  3    24
#>  [49,]    4    8    44  3    24
#>  [50,]    4    9    44  3    24
#>  [51,]    4   10    44  3    24
#>  [52,]    4    1    45  4    21
#>  [53,]    4    2    45  4    21
#>  [54,]    4    3    45  4    21
#>  [55,]    4    4    45  4    21
#>  [56,]    4    5    45  4    21
#>  [57,]    4    6    45  4    21
#>  [58,]    4    7    45  4    21
#>  [59,]    4    8    45  4    21
#>  [60,]    4    9    45  4    21
#>  [61,]    4   10    45  4    21
#>  [62,]    4    1    45  5    27
#>  [63,]    4    2    45  5    27
#>  [64,]    4    3    45  5    27
#>  [65,]    4    4    45  5    27
#>  [66,]    4    5    45  5    27
#>  [67,]    4    6    45  5    27
#>  [68,]    4    7    45  5    27
#>  [69,]    4    8    45  5    27
#>  [70,]    4    9    45  5    27
#>  [71,]    4   10    45  5    27
#>  [72,]    4    1    46  6    21
#>  [73,]    4    2    46  6    21
#>  [74,]    4    3    46  6    21
#>  [75,]    4    4    46  6    21
#>  [76,]    4    5    46  6    21
#>  [77,]    4    6    46  6    21
#>  [78,]    4    7    46  6    21
#>  [79,]    4    8    46  6    21
#>  [80,]    4    9    46  6    21
#>  [81,]    4   10    46  6    21
#>  [82,]    4    1    46  7    27
#>  [83,]    4    2    46  7    27
#>  [84,]    4    3    46  7    27
#>  [85,]    4    4    46  7    27
#>  [86,]    4    5    46  7    27
#>  [87,]    4    6    46  7    27
#>  [88,]    4    7    46  7    27
#>  [89,]    4    8    46  7    27
#>  [90,]    4    9    46  7    27
#>  [91,]    4   10    46  7    27
#>  [92,]    4    1    47  8    21
#>  [93,]    4    2    47  8    21
#>  [94,]    4    3    47  8    21
#>  [95,]    4    4    47  8    21
#>  [96,]    4    5    47  8    21
#>  [97,]    4    6    47  8    21
#>  [98,]    4    7    47  8    21
#>  [99,]    4    8    47  8    21
#> [100,]    4    9    47  8    21
#> [101,]    4   10    47  8    21
#> [102,]    4    1    47  9    27
#> [103,]    4    2    47  9    27
#> [104,]    4    3    47  9    27
#> [105,]    4    4    47  9    27
#> [106,]    4    5    47  9    27
#> [107,]    4    6    47  9    27
#> [108,]    4    7    47  9    27
#> [109,]    4    8    47  9    27
#> [110,]    4    9    47  9    27
#> [111,]    4   10    47  9    27
#> [112,]    4    1    48 10    21
#> [113,]    4    2    48 10    21
#> [114,]    4    3    48 10    21
#> [115,]    4    4    48 10    21
#> [116,]    4    5    48 10    21
#> [117,]    4    6    48 10    21
#> [118,]    4    7    48 10    21
#> [119,]    4    8    48 10    21
#> [120,]    4    9    48 10    21
#> [121,]    4   10    48 10    21
#> [122,]    4    1    48 11    25
#> [123,]    4    2    48 11    25
#> [124,]    4    3    48 11    25
#> [125,]    4    4    48 11    25
#> [126,]    4    5    48 11    25
#> [127,]    4    6    48 11    25
#> [128,]    4    7    48 11    25
#> [129,]    4    8    48 11    25
#> [130,]    4    9    48 11    25
#> [131,]    4   10    48 11    25
#> [132,]    5    1    37  1     6
#> [133,]    6    1    48  1     6
#> [134,]    7    1    43  1     7
#> [135,]    8    1    43  1     7
#> [136,]    9    1    37  1    15
#> [137,]    9    1    37  2    28
#> [138,]   10    1    43  1    10
#> [139,]   10    1    43  2     1
#> [140,]   11    1    43  1     2
#> [141,]   11    1    43  2    20
#> [142,]   11    1    43  3    16
#> [143,]   12    1    37  1     4
#> [144,]   12    1    37  2     8
#> [145,]   13    1    37  1    13
#> [146,]   13    1    38  2     5
#> [147,]   14    1    37  1    13
#> [148,]   14    1    38  2    17
#> [149,]   15    1    38  1    22
#> [150,]   16    1    37  1    19
#> [151,]   16    1    38  2    26