Create a data.frame mapping tests to coverage traces

Extract a matrix used to relate test code to the traces that each test evaluates.

Usage

test_trace_mapping(x)

Arguments

x

A coverage object produced with options(covr.record_tests = TRUE).

Value

A data.frame with one record for each line of code executed, with variables:

test

The index of the test that was executed, reflecting the order in which tests are executed

depth

The call stack depth when the coverage trace was evaluated

i

The index of the expression evaluated by each test. This can be used to recover an order of trace execution for a given test index

trace

The index of the coverage trace that was evaluated

See also

Examples

options(covr.record_tests = TRUE)
pkg_path <- system.file("examplepkg", package = "covtracer")
cov <- covr::package_coverage(pkg_path)
test_trace_mapping(cov)
#>        test depth  i trace
#>   [1,]    1     1  1    24
#>   [2,]    2    37  1    19
#>   [3,]    2    38  2    15
#>   [4,]    2    39  3    23
#>   [5,]    2    40  4    22
#>   [6,]    2    40  5    26
#>   [7,]    2    41  6    22
#>   [8,]    2    41  7    26
#>   [9,]    2    42  8    22
#>  [10,]    2    42  9    26
#>  [11,]    2    43 10    22
#>  [12,]    2    43 11    25
#>  [13,]    3    37  1    23
#>  [14,]    3    38  2    22
#>  [15,]    3    38  3    26
#>  [16,]    3    39  4    22
#>  [17,]    3    39  5    26
#>  [18,]    3    40  6    22
#>  [19,]    3    40  7    26
#>  [20,]    3    41  8    22
#>  [21,]    3    41  9    25
#>  [22,]    4    42  1    19
#>  [23,]    4    43  2    15
#>  [24,]    4    44  3    23
#>  [25,]    4    45  4    22
#>  [26,]    4    45  5    26
#>  [27,]    4    46  6    22
#>  [28,]    4    46  7    26
#>  [29,]    4    47  8    22
#>  [30,]    4    47  9    26
#>  [31,]    4    48 10    22
#>  [32,]    4    48 11    25
#>  [33,]    5    42  1    19
#>  [34,]    5    43  2    15
#>  [35,]    5    44  3    23
#>  [36,]    5    45  4    22
#>  [37,]    5    45  5    26
#>  [38,]    5    46  6    22
#>  [39,]    5    46  7    26
#>  [40,]    5    47  8    22
#>  [41,]    5    47  9    26
#>  [42,]    5    48 10    22
#>  [43,]    5    48 11    25
#>  [44,]    6    42  1    19
#>  [45,]    6    43  2    15
#>  [46,]    6    44  3    23
#>  [47,]    6    45  4    22
#>  [48,]    6    45  5    26
#>  [49,]    6    46  6    22
#>  [50,]    6    46  7    26
#>  [51,]    6    47  8    22
#>  [52,]    6    47  9    26
#>  [53,]    6    48 10    22
#>  [54,]    6    48 11    25
#>  [55,]    7    42  1    19
#>  [56,]    7    43  2    15
#>  [57,]    7    44  3    23
#>  [58,]    7    45  4    22
#>  [59,]    7    45  5    26
#>  [60,]    7    46  6    22
#>  [61,]    7    46  7    26
#>  [62,]    7    47  8    22
#>  [63,]    7    47  9    26
#>  [64,]    7    48 10    22
#>  [65,]    7    48 11    25
#>  [66,]    8    42  1    19
#>  [67,]    8    43  2    15
#>  [68,]    8    44  3    23
#>  [69,]    8    45  4    22
#>  [70,]    8    45  5    26
#>  [71,]    8    46  6    22
#>  [72,]    8    46  7    26
#>  [73,]    8    47  8    22
#>  [74,]    8    47  9    26
#>  [75,]    8    48 10    22
#>  [76,]    8    48 11    25
#>  [77,]    9    42  1    19
#>  [78,]    9    43  2    15
#>  [79,]    9    44  3    23
#>  [80,]    9    45  4    22
#>  [81,]    9    45  5    26
#>  [82,]    9    46  6    22
#>  [83,]    9    46  7    26
#>  [84,]    9    47  8    22
#>  [85,]    9    47  9    26
#>  [86,]    9    48 10    22
#>  [87,]    9    48 11    25
#>  [88,]   10    42  1    19
#>  [89,]   10    43  2    15
#>  [90,]   10    44  3    23
#>  [91,]   10    45  4    22
#>  [92,]   10    45  5    26
#>  [93,]   10    46  6    22
#>  [94,]   10    46  7    26
#>  [95,]   10    47  8    22
#>  [96,]   10    47  9    26
#>  [97,]   10    48 10    22
#>  [98,]   10    48 11    25
#>  [99,]   11    42  1    19
#> [100,]   11    43  2    15
#> [101,]   11    44  3    23
#> [102,]   11    45  4    22
#> [103,]   11    45  5    26
#> [104,]   11    46  6    22
#> [105,]   11    46  7    26
#> [106,]   11    47  8    22
#> [107,]   11    47  9    26
#> [108,]   11    48 10    22
#> [109,]   11    48 11    25
#> [110,]   12    42  1    19
#> [111,]   12    43  2    15
#> [112,]   12    44  3    23
#> [113,]   12    45  4    22
#> [114,]   12    45  5    26
#> [115,]   12    46  6    22
#> [116,]   12    46  7    26
#> [117,]   12    47  8    22
#> [118,]   12    47  9    26
#> [119,]   12    48 10    22
#> [120,]   12    48 11    25
#> [121,]   13    42  1    19
#> [122,]   13    43  2    15
#> [123,]   13    44  3    23
#> [124,]   13    45  4    22
#> [125,]   13    45  5    26
#> [126,]   13    46  6    22
#> [127,]   13    46  7    26
#> [128,]   13    47  8    22
#> [129,]   13    47  9    26
#> [130,]   13    48 10    22
#> [131,]   13    48 11    25
#> [132,]   14    37  1     7
#> [133,]   15    48  1     7
#> [134,]   16    43  1     8
#> [135,]   17    43  1     8
#> [136,]   18    37  1    17
#> [137,]   18    37  2    28
#> [138,]   19    43  1    12
#> [139,]   19    43  2     1
#> [140,]   20    43  1     3
#> [141,]   20    43  2    21
#> [142,]   20    43  3    18
#> [143,]   21    37  1     5
#> [144,]   21    37  2     9
#> [145,]   22    37  1     6
#> [146,]   22    38  2    20
#> [147,]   23    37  1     6
#> [148,]   23    38  2     2
#> [149,]   24    38  1    13
#> [150,]   25    37  1    10
#> [151,]   25    38  2    16