Create a data.frame mapping tests to coverage traces

Extract a matrix used to relate test code to the traces that each test evaluates.

Usage

test_trace_mapping(x)

Arguments

x

A coverage object produced with options(covr.record_tests = TRUE).

Value

A data.frame with one record for each line of code executed, with variables:

test

The index of the test that was executed, reflecting the order in which tests are executed

depth

The call stack depth when the coverage trace was evaluated

i

The index of the expression evaluated by each test. This can be used to recover an order of trace execution for a given test index

trace

The index of the coverage trace that was evaluated

See also

Examples

options(covr.record_tests = TRUE)
pkg_path <- system.file("examplepkg", package = "covtracer")
cov <- covr::package_coverage(pkg_path)
test_trace_mapping(cov)
#>       test depth  i trace
#>  [1,]    1    37  1    15
#>  [2,]    1    38  2    25
#>  [3,]    1    39  3     8
#>  [4,]    1    40  4     1
#>  [5,]    1    40  5    23
#>  [6,]    1    41  6     1
#>  [7,]    1    41  7    23
#>  [8,]    1    42  8     1
#>  [9,]    1    42  9    23
#> [10,]    1    43 10     1
#> [11,]    1    43 11    12
#> [12,]    2    37  1     8
#> [13,]    2    38  2     1
#> [14,]    2    38  3    23
#> [15,]    2    39  4     1
#> [16,]    2    39  5    23
#> [17,]    2    40  6     1
#> [18,]    2    40  7    23
#> [19,]    2    41  8     1
#> [20,]    2    41  9    12
#> [21,]    3    37  1    13
#> [22,]    4    48  1    13
#> [23,]    5    43  1     9
#> [24,]    6    43  1     9
#> [25,]    7    37  1    16
#> [26,]    7    37  2    24
#> [27,]    8    43  1    21
#> [28,]    8    43  2     6
#> [29,]    9    43  1     2
#> [30,]    9    43  2     5
#> [31,]    9    43  3    22
#> [32,]   10    37  1     7
#> [33,]   10    37  2    18
#> [34,]   11    37  1    10
#> [35,]   11    38  2    17
#> [36,]   12    37  1    10
#> [37,]   12    38  2    19
#> [38,]   13    38  1     3