Create a data.frame of coverage trace srcref objects

Extract coverage traces. Traces are the traced lines of code counted when evaluating code coverage, which are used for counting expression evaluation. Each traced is a unique expression within a package's source code.

trace_srcrefs_df(x)

Arguments

x
A package_coverage coverage object, from which the name of the package used is extracted.

Value

A data.frame, where each record it a trace srcref with variables:

name

A character identifier. This will use the names of the elements of a coverage object, which are srcref "keys".

srcref

A srcref object of the trace source code location

See also

Examples

options(covr.record_tests = TRUE)
pkg_path <- system.file("examplepkg", package = "covtracer")
cov <- covr::package_coverage(pkg_path)
trace_srcrefs_df(cov)

#>                                    name                  srcref
#> 1    s3_example.R:20:3:20:8:3:8:207:207  s3_example.R:20:3:20:8
#> 2    r6_example.R:89:9:89:22:9:22:89:89 r6_example.R:89:9:89:22
#> 3    rd_sampler.R:57:3:57:6:3:6:185:185  rd_sampler.R:57:3:57:6
#> 4  s4_example.R:33:3:33:15:3:15:242:242 s4_example.R:33:3:33:15
#> 5  r6_example.R:90:7:90:14:7:14:102:102 r6_example.R:90:7:90:14
#> 6    r6_example.R:54:7:54:26:7:26:66:66 r6_example.R:54:7:54:26
#> 7    r6_example.R:68:7:68:50:7:50:80:80 r6_example.R:68:7:68:50
#> 8  s3_example.R:15:3:15:11:3:11:202:202 s3_example.R:15:3:15:11
#> 9    r6_example.R:55:7:55:24:7:24:67:67 r6_example.R:55:7:55:24
#> 10 s4_example.R:16:3:16:15:3:15:225:225 s4_example.R:16:3:16:15
#> 11   r6_example.R:31:5:31:19:5:19:43:43 r6_example.R:31:5:31:19
#> 12 s3_example.R:10:3:10:30:3:30:197:197 s3_example.R:10:3:10:30
#> 13   r6_example.R:30:5:30:28:5:28:42:42 r6_example.R:30:5:30:28
#> 14     hypotenuse.R:8:3:8:25:3:25:10:10   hypotenuse.R:8:3:8:25
#> 15   r6_example.R:19:5:19:19:5:19:31:31 r6_example.R:19:5:19:19
#> 16   r6_example.R:67:7:67:51:7:51:79:79 r6_example.R:67:7:67:51
#> 17   r6_example.R:66:7:66:23:7:23:78:78 r6_example.R:66:7:66:23
#> 18 r6_example.R:92:7:92:48:7:48:104:104 r6_example.R:92:7:92:48